backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "NutraMap Backend API"
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn>=0.34.0",
    "pymongo>=4.10.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.20",
    "openai>=1.58.1",
    "python-dotenv>=1.0.1",
    "faiss-cpu>=1.9.0.post1",
    "pydantic>=2.10.4",
    "pydantic-core>=2.27.2",
    "email-validator>=2.2.0",
    "numpy>=1.24.4",
    "typesense>=0.21.0",
    "rapidfuzz>=3.11.0",
    "scipy>=1.10.1",
    "modal>=0.65.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = []
